+++ src/renderer/components/dialogs/DeltaDialog.tsx
@@ -156,7 +156,7 @@ export function DeltaDialogFooter(
   if (typeof hide === 'undefined') hide = typeof children === 'undefined'
   return (
     <div
-      style={{ display: hide ? 'none' : 'unset', ...props.style }}
+      style={{ display: hide ? 'none' : 'block', ...props.style }}
       className={classNames(
         Classes.DIALOG_FOOTER,
         'bp3-dialog-footer-border-top'
@@ -293,27 +293,80 @@ export function SmallSelectDialog({
           </RadioGroup>
         </DeltaDialogContent>
       </DeltaDialogBody>
-      <DeltaDialogFooter
-        style={{
-          display: 'flex',
-          justifyContent: 'space-between',
-          marginTop: '0px',
-          padding: '7px 13px 10px 13px',
-        }}
-      >
+      <DeltaDialogFooter style={{ marginTop: '0px', padding: '20px' }}>
+        <DeltaDialogFooterActions>
+          <p
+            className='delta-button primary bold'
+            onClick={() => {
+              onCancel && onCancel()
+              onClose()
+            }}
+          >
+            {tx('cancel')}
+          </p>
+          <p className='delta-button primary bold' onClick={saveAndClose}>
+            {tx('save_desktop')}
+          </p>
+        </DeltaDialogFooterActions>
+      </DeltaDialogFooter>
+    </SmallDialog>
+  )
+}
+
+export function DeltaDialogFooterActions({
+  children,
+  style,
+}: {
+  children: any
+  style?: any
+}) {
+  return (
+    <div
+      style={{ justifyContent: 'flex-end', ...style }}
+      className={Classes.DIALOG_FOOTER_ACTIONS}
+    >
+      {children}
+    </div>
+  )
+}
+
+export function DeltaDialogOkCancelFooter({
+  onCancel,
+  onOk,
+}: {
+  onCancel: () => any
+  onOk: () => any
+}) {
+  const tx = window.static_translate
+
+  return (
+    <DeltaDialogFooter>
+      <DeltaDialogFooterActions>
         <p
-          className='delta-button danger bold'
-          onClick={() => {
-            onCancel && onCancel()
-            onClose()
-          }}
+          className='delta-button primary bold'
+          style={{ marginRight: '10px' }}
+          onClick={onCancel}
         >
           {tx('cancel')}
         </p>
-        <p className='delta-button primary bold' onClick={saveAndClose}>
-          {tx('save_desktop')}
+        <p className={'delta-button bold primary'} onClick={onOk}>
+          {tx('ok')}
         </p>
-      </DeltaDialogFooter>
-    </SmallDialog>
+      </DeltaDialogFooterActions>
+    </DeltaDialogFooter>
+  )
+}
+
+export function DeltaDialogCloseFooter({ onClose }: { onClose: () => any }) {
+  const tx = window.static_translate
+
+  return (
+    <DeltaDialogFooter>
+      <DeltaDialogFooterActions>
+        <p className={'delta-button bold primary'} onClick={onClose}>
+          {tx('close')}
+        </p>
+      </DeltaDialogFooterActions>
+    </DeltaDialogFooter>
   )
 }