+++ src/renderer/components/message/LabeledLink.tsx
@@ -2,7 +2,11 @@ import React, { useContext, useState } from 'react'
 import { ScreenContext } from '../../contexts'
 import { toASCII } from 'punycode'
 import { OpenDialogFunctionType } from '../dialogs/DialogController'
-import { SmallDialog } from '../dialogs/DeltaDialog'
+import {
+  SmallDialog,
+  DeltaDialogFooterActions,
+  DeltaDialogFooter,
+} from '../dialogs/DeltaDialog'
 import { Classes } from '@blueprintjs/core'
 import { DeltaCheckbox } from '../contact/ContactListItem'
 import { getLogger } from '../../../shared/logger'
@@ -94,19 +98,13 @@ function confirmationDialog(
               <i>{domain}</i>
             </div>
           </div>
-          <div className={Classes.DIALOG_FOOTER}>
-            <div
-              className={Classes.DIALOG_FOOTER_ACTIONS}
-              style={{ justifyContent: 'space-between', marginTop: '7px' }}
-            >
-              <p
-                className={`delta-button no-padding bold primary`}
-                onClick={onClose}
-              >
+          <DeltaDialogFooter>
+            <DeltaDialogFooterActions>
+              <p className={`delta-button bold primary`} onClick={onClose}>
                 {tx('no')}
               </p>
               <p
-                className={`delta-button no-padding bold primary`}
+                className={`delta-button bold primary`}
                 onClick={() => {
                   onClose()
                   navigator.clipboard.writeText(target)
@@ -115,7 +113,7 @@ function confirmationDialog(
                 {tx('menu_copy_link_to_clipboard')}
               </p>
               <p
-                className={`delta-button no-padding bold primary`}
+                className={`delta-button bold primary`}
                 onClick={() => {
                   onClose()
                   if (isChecked) {
@@ -127,8 +125,8 @@ function confirmationDialog(
               >
                 {tx('open')}
               </p>
-            </div>
-          </div>
+            </DeltaDialogFooterActions>
+          </DeltaDialogFooter>
         </div>
       </SmallDialog>
     )